suggestions from PR
