+++ Composer/packages/extensions/adaptive-form/src/components/FormRow.tsx
@@ -5,7 +5,6 @@
 import { css, jsx } from '@emotion/core';
 import React from 'react';
 import { FieldProps, UIOptions } from '@bfc/extension';
-import expressionConfig from '@bfc/ui-plugin-expressions';
 
 import { resolvePropSchema } from '../utils';
 
@@ -109,7 +108,6 @@ const FormRow: React.FC<FormRowProps> = props => {
         onBlur={onBlur}
         onChange={onChange(row)}
         onFocus={onFocus}
-        helpURL={expressionConfig.roleSchema?.expression?.helpLink}
       />
     );
   }
