+++ Composer/packages/extensions/adaptive-form/src/components/SchemaField.tsx
@@ -5,7 +5,7 @@ import { jsx, css } from '@emotion/core';
 import React, { useEffect } from 'react';
 import { FieldProps } from '@bfc/extension';
 
-import { getUISchema, resolveFieldWidget, resolveRef, getUiLabel, getUiPlaceholder, getUiDescription } from '../utils';
+import { getUIOptions, resolveFieldWidget, resolveRef, getUiLabel, getUiPlaceholder, getUiDescription } from '../utils';
 import { usePluginConfig } from '../hooks';
 
 import { ErrorMessage } from './ErrorMessage';
@@ -31,13 +31,13 @@ const SchemaField: React.FC<FieldProps> = props => {
     rawErrors,
     hideError,
     onChange,
-    helpURL,
     ...rest
   } = props;
   const pluginConfig = usePluginConfig();
+
   const schema = resolveRef(baseSchema, definitions);
   const uiOptions = {
-    ...getUISchema(schema, pluginConfig.formSchema),
+    ...getUIOptions(schema, pluginConfig.formSchema, pluginConfig.roleSchema[schema.$role || '']),
     ...baseUIOptions,
   };
 
@@ -52,7 +52,7 @@ const SchemaField: React.FC<FieldProps> = props => {
   }, []);
 
   const error = typeof rawErrors === 'string' && (
-    <ErrorMessage error={rawErrors} label={getUiLabel(props)} helpURL={helpURL} />
+    <ErrorMessage error={rawErrors} label={getUiLabel(props)} helpURL={uiOptions.helpLink} />
   );
 
   if (!schema || name.startsWith('$')) {
