+++ Composer/packages/extensions/adaptive-form/src/utils/index.ts
@@ -3,7 +3,7 @@
 export * from './arrayUtils';
 export * from './getAdaptiveType';
 export * from './getOrderedProperties';
-export * from './getUISchema';
+export * from './getUIOptions';
 export * from './getValueType';
 export * from './mergePluginConfigs';
 export * from './resolveBaseSchema';
